all: refactor app logs extracting code to a new package
