+++ api/app_test.go
@@ -42,6 +42,7 @@ import (
 	"github.com/tsuru/tsuru/router/rebuild"
 	"github.com/tsuru/tsuru/router/routertest"
 	"github.com/tsuru/tsuru/service"
+	"github.com/tsuru/tsuru/servicemanager"
 	apiTypes "github.com/tsuru/tsuru/types/api"
 	appTypes "github.com/tsuru/tsuru/types/app"
 	authTypes "github.com/tsuru/tsuru/types/auth"
@@ -4116,13 +4117,13 @@ func (s *S) TestAppLogFollow(c *check.C) {
 		splitted := strings.Split(strings.TrimSpace(recorder.Body.String()), "\n")
 		c.Assert(splitted, check.HasLen, 2)
 		c.Assert(splitted[0], check.Equals, "[]")
-		logs := []app.Applog{}
+		logs := []appTypes.Applog{}
 		logErr = json.Unmarshal([]byte(splitted[1]), &logs)
 		c.Assert(logErr, check.IsNil)
 		c.Assert(logs, check.HasLen, 1)
 		c.Assert(logs[0].Message, check.Equals, "x")
 	}()
-	var listener *app.LogListener
+	var listener appTypes.LogWatcher
 	timeout := time.After(5 * time.Second)
 	for listener == nil {
 		select {
@@ -4135,7 +4136,7 @@ func (s *S) TestAppLogFollow(c *check.C) {
 		}
 		logTracker.Unlock()
 	}
-	err = a.Log("x", "", "")
+	err = servicemanager.AppLog.Add(a.Name, "x", "", "")
 	c.Assert(err, check.IsNil)
 	time.Sleep(500 * time.Millisecond)
 	cancel()
@@ -4165,13 +4166,13 @@ func (s *S) TestAppLogFollowWithFilter(c *check.C) {
 		splitted := strings.Split(strings.TrimSpace(recorder.Body.String()), "\n")
 		c.Assert(splitted, check.HasLen, 2)
 		c.Assert(splitted[0], check.Equals, "[]")
-		logs := []app.Applog{}
+		logs := []appTypes.Applog{}
 		logErr = json.Unmarshal([]byte(splitted[1]), &logs)
 		c.Assert(logErr, check.IsNil)
 		c.Assert(logs, check.HasLen, 1)
 		c.Assert(logs[0].Message, check.Equals, "y")
 	}()
-	var listener *app.LogListener
+	var listener appTypes.LogWatcher
 	timeout := time.After(5 * time.Second)
 	for listener == nil {
 		select {
@@ -4184,9 +4185,9 @@ func (s *S) TestAppLogFollowWithFilter(c *check.C) {
 		}
 		logTracker.Unlock()
 	}
-	err = a.Log("x", "app", "")
+	err = servicemanager.AppLog.Add(a.Name, "x", "app", "")
 	c.Assert(err, check.IsNil)
-	err = a.Log("y", "web", "")
+	err = servicemanager.AppLog.Add(a.Name, "y", "web", "")
 	c.Assert(err, check.IsNil)
 	time.Sleep(500 * time.Millisecond)
 	cancel()
@@ -4215,7 +4216,7 @@ func (s *S) TestAppLogSelectByLines(c *check.C) {
 	err := app.CreateApp(&a, s.user)
 	c.Assert(err, check.IsNil)
 	for i := 0; i < 15; i++ {
-		a.Log(strconv.Itoa(i), "source", "")
+		servicemanager.AppLog.Add(a.Name, strconv.Itoa(i), "source", "")
 	}
 	token := userWithPermission(c, permission.Permission{
 		Scheme:  permission.PermAppReadLog,
@@ -4228,7 +4229,7 @@ func (s *S) TestAppLogSelectByLines(c *check.C) {
 	err = appLog(recorder, request, token)
 	c.Assert(err, check.IsNil)
 	c.Assert(recorder.Code, check.Equals, http.StatusOK)
-	logs := []app.Applog{}
+	logs := []appTypes.Applog{}
 	err = json.Unmarshal(recorder.Body.Bytes(), &logs)
 	c.Assert(err, check.IsNil)
 	c.Assert(logs, check.HasLen, 10)
@@ -4238,8 +4239,8 @@ func (s *S) TestAppLogSelectBySource(c *check.C) {
 	a := app.App{Name: "lost", Platform: "zend", TeamOwner: s.team.Name}
 	err := app.CreateApp(&a, s.user)
 	c.Assert(err, check.IsNil)
-	a.Log("mars log", "mars", "")
-	a.Log("earth log", "earth", "")
+	servicemanager.AppLog.Add(a.Name, "mars log", "mars", "")
+	servicemanager.AppLog.Add(a.Name, "earth log", "earth", "")
 	token := userWithPermission(c, permission.Permission{
 		Scheme:  permission.PermAppReadLog,
 		Context: permission.Context(permTypes.CtxTeam, s.team.Name),
@@ -4252,7 +4253,7 @@ func (s *S) TestAppLogSelectBySource(c *check.C) {
 	err = appLog(recorder, request, token)
 	c.Assert(err, check.IsNil)
 	c.Assert(recorder.Code, check.Equals, http.StatusOK)
-	logs := []app.Applog{}
+	logs := []appTypes.Applog{}
 	err = json.Unmarshal(recorder.Body.Bytes(), &logs)
 	c.Assert(err, check.IsNil)
 	c.Assert(logs, check.HasLen, 1)
@@ -4264,8 +4265,8 @@ func (s *S) TestAppLogSelectByUnit(c *check.C) {
 	a := app.App{Name: "lost", Platform: "zend", TeamOwner: s.team.Name}
 	err := app.CreateApp(&a, s.user)
 	c.Assert(err, check.IsNil)
-	a.Log("mars log", "mars", "prospero")
-	a.Log("earth log", "earth", "caliban")
+	servicemanager.AppLog.Add(a.Name, "mars log", "mars", "prospero")
+	servicemanager.AppLog.Add(a.Name, "earth log", "earth", "caliban")
 	token := userWithPermission(c, permission.Permission{
 		Scheme:  permission.PermAppReadLog,
 		Context: permission.Context(permTypes.CtxTeam, s.team.Name),
@@ -4278,7 +4279,7 @@ func (s *S) TestAppLogSelectByUnit(c *check.C) {
 	err = appLog(recorder, request, token)
 	c.Assert(err, check.IsNil)
 	c.Assert(recorder.Code, check.Equals, http.StatusOK)
-	logs := []app.Applog{}
+	logs := []appTypes.Applog{}
 	err = json.Unmarshal(recorder.Body.Bytes(), &logs)
 	c.Assert(err, check.IsNil)
 	c.Assert(logs, check.HasLen, 1)
@@ -4295,7 +4296,7 @@ func (s *S) TestAppLogSelectByLinesShouldReturnTheLastestEntries(c *check.C) {
 	coll, err := s.logConn.CreateAppLogCollection(a.Name)
 	c.Assert(err, check.IsNil)
 	for i := 0; i < 15; i++ {
-		l := app.Applog{
+		l := appTypes.Applog{
 			Date:    now.Add(time.Duration(i) * time.Hour),
 			Message: strconv.Itoa(i),
 			Source:  "source",
@@ -4315,7 +4316,7 @@ func (s *S) TestAppLogSelectByLinesShouldReturnTheLastestEntries(c *check.C) {
 	err = appLog(recorder, request, token)
 	c.Assert(err, check.IsNil)
 	c.Assert(recorder.Code, check.Equals, http.StatusOK)
-	var logs []app.Applog
+	var logs []appTypes.Applog
 	err = json.Unmarshal(recorder.Body.Bytes(), &logs)
 	c.Assert(err, check.IsNil)
 	c.Assert(logs, check.HasLen, 3)
@@ -4328,15 +4329,15 @@ func (s *S) TestAppLogShouldReturnLogByApp(c *check.C) {
 	app1 := app.App{Name: "app1", Platform: "zend", TeamOwner: s.team.Name}
 	err := app.CreateApp(&app1, s.user)
 	c.Assert(err, check.IsNil)
-	app1.Log("app1 log", "source", "")
+	servicemanager.AppLog.Add(app1.Name, "app1 log", "source", "")
 	app2 := app.App{Name: "app2", Platform: "zend", TeamOwner: s.team.Name}
 	err = app.CreateApp(&app2, s.user)
 	c.Assert(err, check.IsNil)
-	app2.Log("app2 log", "sourc ", "")
+	servicemanager.AppLog.Add(app2.Name, "app2 log", "sourc ", "")
 	app3 := app.App{Name: "app3", Platform: "zend", TeamOwner: s.team.Name}
 	err = app.CreateApp(&app3, s.user)
 	c.Assert(err, check.IsNil)
-	app3.Log("app3 log", "tsuru", "")
+	servicemanager.AppLog.Add(app3.Name, "app3 log", "tsuru", "")
 	token := userWithPermission(c, permission.Permission{
 		Scheme:  permission.PermAppReadLog,
 		Context: permission.Context(permTypes.CtxTeam, s.team.Name),
@@ -4349,7 +4350,7 @@ func (s *S) TestAppLogShouldReturnLogByApp(c *check.C) {
 	err = appLog(recorder, request, token)
 	c.Assert(err, check.IsNil)
 	c.Assert(recorder.Code, check.Equals, http.StatusOK)
-	logs := []app.Applog{}
+	logs := []appTypes.Applog{}
 	err = json.Unmarshal(recorder.Body.Bytes(), &logs)
 	c.Assert(err, check.IsNil)
 	var logged bool
@@ -5471,7 +5472,7 @@ func (s *S) TestSleepHandlerReturns403IfTheUserDoesNotHaveAccessToTheApp(c *chec
 	c.Assert(e.Code, check.Equals, http.StatusForbidden)
 }
 
-type LogList []app.Applog
+type LogList []appTypes.Applog
 
 func (l LogList) Len() int           { return len(l) }
 func (l LogList) Swap(i, j int)      { l[i], l[j] = l[j], l[i] }
@@ -5521,7 +5522,7 @@ func (s *S) TestAddLog(c *check.C) {
 		"mysource",
 		"mysource",
 	}
-	logs, err := a.LastLogs(5, app.Applog{})
+	logs, err := a.LastLogs(5, appTypes.Applog{})
 	c.Assert(err, check.IsNil)
 	got := make([]string, len(logs))
 	gotSource := make([]string, len(logs))
@@ -6285,17 +6286,17 @@ func (s *S) TestFollowLogs(c *check.C) {
 	enc := &fakeEncoder{
 		done: make(chan struct{}),
 	}
-	l, err := app.NewLogListener(&a, app.Applog{})
+	l, err := servicemanager.AppLog.Watch(a.Name, "", "")
 	c.Assert(err, check.IsNil)
 	go func() {
-		err = a.Log("xyz", "", "")
+		err = servicemanager.AppLog.Add(a.Name, "xyz", "", "")
 		c.Assert(err, check.IsNil)
 		<-enc.done
 		cancel()
 	}()
 	err = followLogs(ctx, l, enc)
 	c.Assert(err, check.IsNil)
-	msgSlice, ok := enc.msg.([]app.Applog)
+	msgSlice, ok := enc.msg.([]appTypes.Applog)
 	c.Assert(ok, check.Equals, true)
 	c.Assert(msgSlice, check.HasLen, 1)
 	c.Assert(msgSlice[0].Message, check.Equals, "xyz")
@@ -6313,7 +6314,7 @@ func (s *S) TestFollowLogsTimeout(c *check.C) {
 	enc := &fakeEncoder{
 		done: make(chan struct{}),
 	}
-	l, err := app.NewLogListener(&a, app.Applog{})
+	l, err := servicemanager.AppLog.Watch(a.Name, "", "")
 	c.Assert(err, check.IsNil)
 	err = followLogs(ctx, l, enc)
 	c.Assert(err, check.ErrorMatches, `.*timeout.*`)
