+++ api/logtracker_test.go
@@ -8,26 +8,27 @@ import (
 	"context"
 	"time"
 
-	"github.com/tsuru/tsuru/app"
+	"github.com/tsuru/tsuru/servicemanager"
+	appTypes "github.com/tsuru/tsuru/types/app"
 	check "gopkg.in/check.v1"
 )
 
 func (s *S) TestLogStreamTrackerAddRemove(c *check.C) {
 	c.Assert(logTracker.conn, check.HasLen, 0)
-	l := app.LogListener{}
-	logTracker.add(&l)
+	var l appTypes.LogWatcher = nil
+	logTracker.add(l)
 	c.Assert(logTracker.conn, check.HasLen, 1)
-	logTracker.remove(&l)
+	logTracker.remove(l)
 	c.Assert(logTracker.conn, check.HasLen, 0)
 }
 
 func (s *S) TestLogStreamTrackerShutdown(c *check.C) {
-	l, err := app.NewLogListener(&app.App{Name: "myapp"}, app.Applog{})
+	l, err := servicemanager.AppLog.Watch("myapp", "", "")
 	c.Assert(err, check.IsNil)
 	logTracker.add(l)
 	logTracker.Shutdown(context.Background())
 	select {
-	case <-l.ListenChan():
+	case <-l.Chan():
 	case <-time.After(5 * time.Second):
 		c.Fatal("timed out waiting for channel to close")
 	}
