+++ app/app.go
@@ -242,16 +242,6 @@ func (app *App) MarshalJSON() ([]byte, error) {
 	return json.Marshal(&result)
 }
 
-// Applog represents a log entry.
-type Applog struct {
-	MongoID bson.ObjectId `bson:"_id,omitempty" json:"-"`
-	Date    time.Time
-	Message string
-	Source  string
-	AppName string
-	Unit    string
-}
-
 // AcquireApplicationLock acquires an application lock by setting the lock
 // field in the database.  This method is already called by a connection
 // middleware on requests with :app or :appname params that have side-effects.
@@ -1156,7 +1146,7 @@ func (app *App) setEnv(env bind.EnvVar) {
 	}
 	app.Env[env.Name] = env
 	if env.Public {
-		app.Log(fmt.Sprintf("setting env %s with value %s", env.Name, env.Value), "tsuru", "api")
+		servicemanager.AppLog.Add(app.Name, fmt.Sprintf("setting env %s with value %s", env.Name, env.Value), "tsuru", "api")
 	}
 }
 
@@ -1295,10 +1285,10 @@ func (app *App) Run(cmd string, w io.Writer, args provision.RunArgs) error {
 	if !args.Isolated && !app.available() {
 		return errors.New("App must be available to run non-isolated commands")
 	}
-	app.Log(fmt.Sprintf("running '%s'", cmd), "tsuru", "api")
-	logWriter := LogWriter{App: app, Source: "app-run"}
+	logWriter := LogWriter{AppName: app.Name, Source: "app-run"}
 	logWriter.Async()
 	defer logWriter.Close()
+	logWriter.Write([]byte(fmt.Sprintf("running '%s'", cmd)))
 	return app.run(cmd, io.MultiWriter(w, &logWriter), args)
 }
 
@@ -1749,45 +1739,13 @@ func (app *App) RemoveInstance(removeArgs bind.RemoveInstanceArgs) error {
 	return nil
 }
 
-// Log adds a log message to the app. Specifying a good source is good so the
-// user can filter where the message come from.
-func (app *App) Log(message, source, unit string) error {
-	messages := strings.Split(message, "\n")
-	logs := make([]interface{}, 0, len(messages))
-	for _, msg := range messages {
-		if msg != "" {
-			l := Applog{
-				Date:    time.Now().In(time.UTC),
-				Message: msg,
-				Source:  source,
-				AppName: app.Name,
-				Unit:    unit,
-			}
-			logs = append(logs, l)
-		}
-	}
-	if len(logs) > 0 {
-		conn, err := db.LogConn()
-		if err != nil {
-			return err
-		}
-		defer conn.Close()
-		coll, err := conn.CreateAppLogCollection(app.Name)
-		if err != nil {
-			return err
-		}
-		return coll.Insert(logs...)
-	}
-	return nil
-}
-
 // LastLogs returns a list of the last `lines` log of the app, matching the
 // fields in the log instance received as an example.
-func (app *App) LastLogs(lines int, filterLog Applog) ([]Applog, error) {
+func (app *App) LastLogs(lines int, filterLog appTypes.Applog) ([]appTypes.Applog, error) {
 	return app.lastLogs(lines, filterLog, false)
 }
 
-func (app *App) lastLogs(lines int, filterLog Applog, invertFilter bool) ([]Applog, error) {
+func (app *App) lastLogs(lines int, filterLog appTypes.Applog, invertFilter bool) ([]appTypes.Applog, error) {
 	prov, err := app.getProvisioner()
 	if err != nil {
 		return nil, err
@@ -1804,33 +1762,13 @@ func (app *App) lastLogs(lines int, filterLog Applog, invertFilter bool) ([]Appl
 			return nil, errors.New(doc)
 		}
 	}
-	conn, err := db.LogConn()
-	if err != nil {
-		return nil, err
-	}
-	defer conn.Close()
-	logs := []Applog{}
-	q := bson.M{}
-	if filterLog.Source != "" {
-		q["source"] = filterLog.Source
-	}
-	if filterLog.Unit != "" {
-		q["unit"] = filterLog.Unit
-	}
-	if invertFilter {
-		for k, v := range q {
-			q[k] = bson.M{"$ne": v}
-		}
-	}
-	err = conn.AppLogCollection(app.Name).Find(q).Sort("-$natural").Limit(lines).All(&logs)
-	if err != nil {
-		return nil, err
-	}
-	l := len(logs)
-	for i := 0; i < l/2; i++ {
-		logs[i], logs[l-1-i] = logs[l-1-i], logs[i]
-	}
-	return logs, nil
+	return servicemanager.AppLog.List(appTypes.ListLogArgs{
+		AppName:       app.Name,
+		InvertFilters: invertFilter,
+		Limit:         lines,
+		Source:        filterLog.Source,
+		Unit:          filterLog.Unit,
+	})
 }
 
 type Filter struct {
@@ -2490,7 +2428,7 @@ func (app *App) Unlock() {
 }
 
 func (app *App) withLogWriter(w io.Writer) io.Writer {
-	logWriter := &LogWriter{App: app}
+	logWriter := &LogWriter{AppName: app.Name}
 	if w != nil {
 		w = io.MultiWriter(w, logWriter)
 	} else {
