+++ app/deploy.go
@@ -24,6 +24,7 @@ import (
 	"github.com/tsuru/tsuru/provision"
 	"github.com/tsuru/tsuru/router/rebuild"
 	"github.com/tsuru/tsuru/set"
+	appTypes "github.com/tsuru/tsuru/types/app"
 	permTypes "github.com/tsuru/tsuru/types/permission"
 )
 
@@ -208,7 +209,7 @@ func Build(opts DeployOptions) (string, error) {
 	if opts.Event == nil {
 		return "", errors.Errorf("missing event in build opts")
 	}
-	logWriter := LogWriter{App: opts.App}
+	logWriter := LogWriter{AppName: opts.App.Name}
 	logWriter.Async()
 	defer logWriter.Close()
 	opts.Event.SetLogWriter(io.MultiWriter(&tsuruIo.NoErrorWriter{Writer: opts.OutputStream}, &logWriter))
@@ -232,7 +233,7 @@ func Build(opts DeployOptions) (string, error) {
 
 type errorWithLog struct {
 	err  error
-	logs []Applog
+	logs []appTypes.Applog
 }
 
 func (e *errorWithLog) Cause() error {
@@ -269,7 +270,7 @@ func Deploy(opts DeployOptions) (string, error) {
 		}
 		opts.Image = imageName
 	}
-	logWriter := LogWriter{App: opts.App}
+	logWriter := LogWriter{AppName: opts.App.Name}
 	logWriter.Async()
 	defer logWriter.Close()
 	opts.Event.SetLogWriter(io.MultiWriter(&tsuruIo.NoErrorWriter{Writer: opts.OutputStream}, &logWriter))
@@ -280,9 +281,9 @@ func Deploy(opts DeployOptions) (string, error) {
 		log.Errorf("WARNING: unable to ensure quota is up-to-date after deploy: %v", quotaErr)
 	}
 	if err != nil {
-		var logLines []Applog
+		var logLines []appTypes.Applog
 		if provision.IsStartupError(err) {
-			logLines, _ = opts.App.lastLogs(10, Applog{
+			logLines, _ = opts.App.lastLogs(10, appTypes.Applog{
 				Source: "tsuru",
 			}, true)
 		}
