+++ applog/dispatcher.go
@@ -1,25 +1,23 @@
-// Copyright 2013 tsuru authors. All rights reserved.
+// Copyright 2019 tsuru authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package app
+package applog
 
 import (
 	"context"
 	"fmt"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
 
-	"github.com/globalsign/mgo"
-	"github.com/globalsign/mgo/bson"
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/tsuru/config"
 	"github.com/tsuru/tsuru/api/shutdown"
 	"github.com/tsuru/tsuru/db"
 	"github.com/tsuru/tsuru/log"
+	appTypes "github.com/tsuru/tsuru/types/app"
 	"golang.org/x/time/rate"
 )
 
@@ -111,122 +109,6 @@ func init() {
 	prometheus.MustRegister(logsAppTailEntries)
 }
 
-type LogListener struct {
-	c       <-chan Applog
-	logConn *db.LogStorage
-	quit    chan struct{}
-}
-
-func isCappedPositionLost(err error) bool {
-	if err == nil {
-		return false
-	}
-	return strings.Contains(err.Error(), "CappedPositionLost")
-}
-
-func isSessionClosed(r interface{}) bool {
-	return fmt.Sprintf("%v", r) == "Session already closed"
-}
-
-func NewLogListener(a *App, filterLog Applog) (*LogListener, error) {
-	conn, err := db.LogConn()
-	if err != nil {
-		return nil, err
-	}
-	c := make(chan Applog, 10)
-	quit := make(chan struct{})
-	coll := conn.AppLogCollection(a.Name)
-	var lastLog Applog
-	err = coll.Find(nil).Sort("-_id").Limit(1).One(&lastLog)
-	if err == mgo.ErrNotFound {
-		// Tail cursors do not work correctly if the collection is empty (the
-		// Next() call wouldn't block). So if the collection is empty we insert
-		// the very first log line in it. This is quite rare in the real world
-		// though so the impact of this extra log message is really small.
-		err = a.Log("Logs initialization", "tsuru", "")
-		if err != nil {
-			conn.Close()
-			return nil, err
-		}
-		err = coll.Find(nil).Sort("-_id").Limit(1).One(&lastLog)
-	}
-	if err != nil {
-		conn.Close()
-		return nil, err
-	}
-	lastId := lastLog.MongoID
-	mkQuery := func() bson.M {
-		m := bson.M{
-			"_id": bson.M{"$gt": lastId},
-		}
-		if filterLog.Source != "" {
-			m["source"] = filterLog.Source
-		}
-		if filterLog.Unit != "" {
-			m["unit"] = filterLog.Unit
-		}
-		return m
-	}
-	query := coll.Find(mkQuery())
-	tailTimeout := 10 * time.Second
-	iter := query.Sort("$natural").Tail(tailTimeout)
-	tailCountMetric := logsAppTail.WithLabelValues(a.Name)
-	entriesMetric := logsAppTailEntries.WithLabelValues(a.Name)
-	go func() {
-		tailCountMetric.Inc()
-		defer tailCountMetric.Dec()
-		defer close(c)
-		defer func() {
-			if r := recover(); r != nil {
-				if isSessionClosed(r) {
-					return
-				}
-				panic(err)
-			}
-		}()
-		for {
-			var applog Applog
-			for iter.Next(&applog) {
-				lastId = applog.MongoID
-				select {
-				case c <- applog:
-					entriesMetric.Inc()
-				case <-quit:
-					iter.Close()
-					return
-				}
-			}
-			if iter.Timeout() {
-				continue
-			}
-			if err := iter.Err(); err != nil {
-				if !isCappedPositionLost(err) {
-					log.Errorf("error tailing logs: %v", err)
-					iter.Close()
-					return
-				}
-			}
-			iter.Close()
-			query = coll.Find(mkQuery())
-			iter = query.Sort("$natural").Tail(tailTimeout)
-		}
-	}()
-	l := LogListener{c: c, logConn: conn, quit: quit}
-	return &l, nil
-}
-
-func (l *LogListener) ListenChan() <-chan Applog {
-	return l.c
-}
-
-func (l *LogListener) Close() {
-	l.logConn.Close()
-	if l.quit != nil {
-		close(l.quit)
-		l.quit = nil
-	}
-}
-
 type LogDispatcher struct {
 	mu             sync.RWMutex
 	dispatchers    map[string]*appLogDispatcher
@@ -236,11 +118,11 @@ type LogDispatcher struct {
 }
 
 type msgWithTS struct {
-	msg        *Applog
+	msg        *appTypes.Applog
 	arriveTime time.Time
 }
 
-func NewlogDispatcher(chanSize int) *LogDispatcher {
+func newlogDispatcher(chanSize int) *LogDispatcher {
 	d := &LogDispatcher{
 		dispatchers:    make(map[string]*appLogDispatcher),
 		msgCh:          make(chan *msgWithTS, chanSize),
@@ -252,7 +134,7 @@ func NewlogDispatcher(chanSize int) *LogDispatcher {
 	return d
 }
 
-func (d *LogDispatcher) getMessageDispatcher(msg *Applog) *appLogDispatcher {
+func (d *LogDispatcher) getMessageDispatcher(msg *appTypes.Applog) *appLogDispatcher {
 	appName := msg.AppName
 	d.mu.RLock()
 	appD, ok := d.dispatchers[appName]
@@ -283,7 +165,7 @@ func (d *LogDispatcher) runWriter() {
 	}
 }
 
-func (d *LogDispatcher) Send(msg *Applog) error {
+func (d *LogDispatcher) Send(msg *appTypes.Applog) error {
 	if atomic.LoadInt32(&d.shuttingDown) == 1 {
 		return errors.New("log dispatcher is shutting down")
 	}
@@ -406,8 +288,8 @@ func (p *bulkProcessor) stopWait() {
 	<-p.finished
 }
 
-func (p *bulkProcessor) rateLimitWarning(rateLimit int) *Applog {
-	return &Applog{
+func (p *bulkProcessor) rateLimitWarning(rateLimit int) *appTypes.Applog {
+	return &appTypes.Applog{
 		AppName: p.appName,
 		Date:    time.Now(),
 		Message: fmt.Sprintf("Log messages dropped due to exceeded rate limit. Limit: %v logs/s.", rateLimit),
@@ -416,8 +298,8 @@ func (p *bulkProcessor) rateLimitWarning(rateLimit int) *Applog {
 	}
 }
 
-func (p *bulkProcessor) globalRateLimitWarning() *Applog {
-	return &Applog{
+func (p *bulkProcessor) globalRateLimitWarning() *appTypes.Applog {
+	return &appTypes.Applog{
 		AppName: p.appName,
 		Date:    time.Now(),
 		Message: fmt.Sprintf("Log messages dropped due to exceeded global rate limit. Global Limit: %v logs/s.", globalRateLimiter.Limit()),
@@ -426,8 +308,8 @@ func (p *bulkProcessor) globalRateLimitWarning() *Applog {
 	}
 }
 
-func (p *bulkProcessor) flushErrorMessage(err error) *Applog {
-	return &Applog{
+func (p *bulkProcessor) flushErrorMessage(err error) *appTypes.Applog {
+	return &appTypes.Applog{
 		AppName: p.appName,
 		Date:    time.Now(),
 		Message: fmt.Sprintf("Log messages dropped due to mongodb insert error: %v", err),
@@ -483,7 +365,7 @@ func (p *bulkProcessor) run() {
 				logsDropped.Inc()
 				if time.Since(lastRateNotice) > rateLimitWarningInterval {
 					lastRateNotice = time.Now()
-					var warning *Applog
+					var warning *appTypes.Applog
 					if globalAllow {
 						warning = p.rateLimitWarning(rateLimiter.Burst())
 					} else {
